+++ src/app/game/player-entry/player-entry.component.scss
