+++ src/app/game/player-entry/team/player-card/player-card.component.scss
