+++ src/app/game/player-entry/team/team.component.html
@@ -9,12 +9,6 @@
         </div>
     </div>
     <div class="player-addition-container">
-        <form [formGroup]="playerForm" class="form">
-            <div class="input-group">
-                <label class="label" for="code-name">Code name</label>
-                <input type="text" id="code-name" class="input" formControlName="codeName">
-            </div>
-        </form>
-        <button [disabled]="playerForm.invalid" class="btn" (click)="addPlayer()">Add player</button>
+        
     </div>
 </div>
