+++ src/app/game/player-entry/team/team.component.scss
@@ -31,6 +31,10 @@
                 .label {
                     color: #a3a3a3;
                 }
+
+                .error-label {
+                    color: #ff0000;
+                }
     
                 .input {
                     width: 100%;
